wsnm: record offsets are shorts, and notes
A byte array subscripted by a long: the pointer and the long both want
HL, and neither a rule nor a cost can put one of them elsewhere.  That
is the register pressure case rather than a missing form.

The source is a stretch either way.  array[long] is legal C when the
value is in range, but on a machine with 16-bit pointers the long half
can never reach an address - the arithmetic is done wide and thrown
away.  Where the offsets can be shown to fit, narrowing them in the
source is the cheaper answer: wsnm's record offsets index one malloc,
and that malloc has to fit the address space, so they are shorts now.

One marker left, in the same function, indexing by a whole-file offset.
Narrowing that one means deciding what the host tool should do with a
file bigger than the target could ever hold, which is a different
question.

PROGRESS.md: sizes remeasured, the valgrind sweep listed with what it
is for, and a bug class added for a field read before it is written -
the one whose symptom is that identical trees compile differently
depending on what came before them.

+++ tools/wsnm.c
@@ -2343,8 +2343,17 @@ char *name;
                 if (endoff > bsssize) bsssize = endoff;
             }
         } else if (rectype == HT_RELOC && (rflag || gflag || dflag) && in_text_psect) {
-            long roff = off;
-            long rend = off + reclen;
+            /*
+             * Offsets into filebuf, which is one malloc: on a machine
+             * with 16-bit pointers the buffer cannot be bigger than
+             * the address space, so these fit a short by construction.
+             * Declared long, every subscript below became a 32-bit sum
+             * truncated to 16 bits to make an address - and the add of
+             * a short to a long wants both halves in HL, which is the
+             * one shape pass2 has no answer for.
+             */
+            unsigned short roff = off;
+            unsigned short rend = off + reclen;
 
             while (roff < rend) {
                 unsigned short reloff;
